+++ netket/sampler/rules/continuous_gaussian.py
@@ -17,7 +17,7 @@ import numpy as np
 
 from flax import struct
 
-from ..metropolis import MetropolisRule
+from .base import MetropolisRule
 
 
 @struct.dataclass
