+++ netket/sampler/rules/hamiltonian_numpy.py
@@ -21,8 +21,7 @@ from flax import struct
 
 from netket.operator import AbstractOperator
 
-
-from ..metropolis import MetropolisRule
+from .base import MetropolisRule
 
 
 @struct.dataclass
